Combine Issue and PR command
