+++ app/src/main/kotlin/me/ghostbear/kumaslash/commands/github/IssueCommand.kt
@@ -2,6 +2,7 @@ package me.ghostbear.kumaslash.commands.github
 
 import com.haroldadmin.opengraphKt.getOpenGraphTags
 import dev.kord.core.behavior.interaction.response.respond
+import dev.kord.core.entity.interaction.InteractionCommand
 import dev.kord.core.event.interaction.GuildChatInputCommandInteractionCreateEvent
 import dev.kord.rest.builder.interaction.string
 import java.net.URL
@@ -10,9 +11,17 @@ import me.ghostbear.core.OnGuildChatInputCommandInteractionCreateEvent
 import me.ghostbear.core.SubSlashCommand
 import me.ghostbear.core.SubSlashCommandConfig
 
+internal const val OWNER = "tachiyomiorg"
+internal const val ARG_REPOSITORY = "repository"
+internal const val ARG_NUMBER = "number"
+
+fun InteractionCommand.getArguments(): Pair<String, String?> {
+    return strings[ARG_REPOSITORY]!! to strings[ARG_NUMBER]
+}
+
 class IssueCommand : SubSlashCommand(), OnGuildChatInputCommandInteractionCreateEvent {
     override val name: String = "issue"
-    override val description: String = "Get issues from GitHub"
+    override val description: String = "Get issues and pull requests from GitHub"
     override val config: SubSlashCommandConfig = {
         string(ARG_REPOSITORY, "GitHub repository") {
             required = true
@@ -22,7 +31,7 @@ class IssueCommand : SubSlashCommand(), OnGuildChatInputCommandInteractionCreate
             choice("Tachiyomi 1.x", "tachiyomi-1.x")
             choice("Tachiyomi Extensions 1.x", "tachiyomi-extensions-1.x")
         }
-        string(ARG_NUMBER, "GitHub issue ID") {
+        string(ARG_NUMBER, "GitHub issue or pull request ID") {
             required = true
         }
     }
